Fix Build (amd_darwin / arm_darwin): cast `fixed_keys_size` to `UInt64`
`row_sizes` is `PaddedPODArray<UInt64>`. On Darwin, `size_t` is `unsigned long`
while `UInt64` is `unsigned long long`. They have identical width but distinct
types, so the value-overload `PODArray::assign(size_t n, const T & x)` requires
an implicit conversion, while the iterator-overload `assign(It1, It2)` is a
perfect match via template deduction. The iterator overload then tries to
dereference an integer ("indirection requires pointer operand").

On Linux both are `unsigned long`, so the value overload wins without
conversion, and the call compiles. Explicit cast resolves the ambiguity on
every platform.

+++ src/Common/ColumnsHashing.h
@@ -632,7 +632,7 @@ struct HashMethodABSerialized
         }
         else
         {
-            row_sizes.assign(rows, fixed_keys_size);
+            row_sizes.assign(rows, static_cast<UInt64>(fixed_keys_size));
         }
 
         size_t total_size = 0;
